# docs/openapi.yaml
openapi: 3.1.0
info:
  title: ZIAY API
  version: 1.0.0
  description: |
    Comercio Conversacional + Atribución Inteligente.

    Multi-tenant conversational-commerce platform with agent-driven checkout
    (AP2 / UCP / ACP / MCP protocol trinity), ad attribution, compliance
    (Ley 1581 / Ley 2573 / Ley 1480 / DIAN), and a wallet/commission layer
    for traffickers + marketplaces.

    All tenant-scoped endpoints require a session cookie (`sessionAuth`)
    unless marked otherwise. UCP / ACP / webhook endpoints use `bearerAuth`
    or are public.

  contact:
    email: api@ziay.co
  license:
    name: Proprietary
servers:
  - url: https://ziay.co
    description: Production
  - url: http://localhost:3000
    description: Development
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: NextAuth v4 session cookie (set by /api/auth/[...nextauth]/callback/credentials).
    bearerAuth:
      type: http
      scheme: bearer
      description: Signed W3C Verifiable Credential or bearer token (UCP/AP2/ACP).
    cronAuth:
      type: apiKey
      in: header
      name: Authorization
      description: 'Bearer $CRON_SECRET — shared secret for /api/compliance/retention/cron.'
    webhookSignature:
      type: apiKey
      in: header
      name: X-Hub-Signature-256
      description: HMAC-SHA256 signature from the gateway (Meta/Stripe/Wompi/etc.).
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        details:
          type: object
          additionalProperties: true
      required: [error]
...
  responses:
...
    Forbidden:
      description: Tenant access denied.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
security:
  - sessionAuth: []
paths:
...
  /api/ads/{id}:
    patch:
      summary: Kill / pause / resume / scale an ad
      description: Pushes the action to the ad platform (simulated in dev).
      security:
        - sessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                action:
                  type: string
                  enum: [pause, kill, resume, scale]
                reason:
                  type: string
      responses:
        '200':
          description: Updated ad + action taken.
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/ads/import:
    post:
      summary: Import ad spend from platform
      description: Importa spend/impressions/clicks de Google/TikTok Ads. Rate-limited (20 req/min).
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                tenantId:
                  type: string
                platform:
                  type: string
                  enum: [google, tiktok]
                dateStart:
                  type: string
                  format: date
                dateEnd:
                  type: string
                  format: date
      responses:
        '200':
          description: 'Import summary: campaignsFetched, adsProcessed, spendUpserted.'
...